// bidly/apps/server/src/app/open-api/bid/types/bid-foreign.type.ts
/**
 * 입찰공고 (외자)
 * @author 최강훈 <ganghun@lepisode.team>
 */
export type Bid_Foreign = {
  bidNtceNo: string;
  bidNtceOrd: string;
  reNtceYn: string;
  rgstTyNm: string;
  ntceKindNm: string;
  intrbidYn: string;
  bidNtceDt: string;
  refNo: string;
  bidNtceNm: string;
  ntceInsttCd: string;
  ntceInsttNm: string;
  dminsttCd: string;
  dminsttNm: string;
  bidMethdNm: string;
  cntrctCnclsMthdNm: string;
  ntceInsttOfclNm: string;
  ntceInsttOfclTelNo: string;
  ntceInsttOfclEmailAdrs: string;
  exctvNm: string;
  bidQlfctRgstDt: string;
  cmmnSpldmdAgrmntRcptdocMethd: string;
  cmmnSpldmdAgrmntClseDt: string;
  cmmnSpldmdCorpRgnLmtYn: string;
  bidBeginDt: string;
  bidClseDt: string;
  opengDt: string;
  ntceSpecDocUrl1: string;
  ntceSpecDocUrl2: string;
  ntceSpecDocUrl3: string;
  ntceSpecDocUrl4: string;
  ntceSpecDocUrl5: string;
  ntceSpecDocUrl6: string;
  ntceSpecDocUrl7: string;
  ntceSpecDocUrl8: string;
  ntceSpecDocUrl9: string;
  ntceSpecDocUrl10: string;
  ntceSpecFileNm1: string;
  ntceSpecFileNm2: string;
  ntceSpecFileNm3: string;
  ntceSpecFileNm4: string;
  ntceSpecFileNm5: string;
  ntceSpecFileNm6: string;
  ntceSpecFileNm7: string;
  ntceSpecFileNm8: string;
  ntceSpecFileNm9: string;
  ntceSpecFileNm10: string;
  rbidPermsnYn: string;
  prdctClsfcLmtYn: string;
  mnfctYn: string;
  prearngPrceDcsnMthdNm: string;
  totPrdprcNum: string;
  drwtPrdprcNum: string;
  asignBdgtAmt: string;
  presmptPrce: string;
  opengPlce: string;
  bidNtceDtlUrl: string;
  bidNtceUrl: string;
  bidPrtcptFeePaymntYn: string;
  bidPrtcptFee: string;
  bidGrntymnyPaymntYn: string;
  crdtrNm: string;
  prdctSno: string;
  dtilPrdctClsfcNo: string;
  dtilPrdctClsfcNoNm: string;
  prdctSpecNm: string;
  prdctQty: string;
  prdctUnit: string;
  prdctUprc: string;
  dlvrTmlmtDt: string;
  dlvrDaynum: string;
  dlvryCndtnNm: string;
  purchsObjPrdctList: string;
  untyNtceNo: string;
  cmmnSpldmdMethdCd: string;
  cmmnSpldmdMethdNm: string;
  stdNtceDocUrl: string;
  brffcBidprcPermsnYn: string;
  dsgntCmptYn: string;
  rsrvtnPrceReMkngMthdNm: string;
  arsltApplDocRcptMthdNm: string;
  arsltApplDocRcptDt: string;
  orderPlanUntyNo: string;
  sucsfbidLwltRate: string;
  rgstDt: string;
  bfSpecRgstNo: string;
  sucsfbidMthdCd: string;
  sucsfbidMthdNm: string;
  chgDt: string;
  dminsttOfclEmailAdrs: string;
  indstrytyLmtYn: string;
  chgNtceRsn: string;
  rbidOpengDt: string;
  bidWgrnteeRcptClseDt: string;
};